[refactor] #3175: Move `iroha_wasm` derive macros parameters parsing into a separate module
